Move StableCoroutineScope to circuit internal
